Move new implementation to its own folder
